+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Set environment tag for native layer.
+
 ## 0.5.1
 
 - Add meta package to dependencies in `pubspec.yaml`.